+++ contracts/eosiolib/varint.hpp
@@ -102,7 +102,7 @@ struct signed_int {
 
     template<typename DataStream>
     friend DataStream& operator << ( DataStream& ds, const signed_int& v ){
-      uint32_t val = (v.value<<1) ^ (v.value>>31);
+      uint32_t val = uint32_t((v.value<<1) ^ (v.value>>31));
       do {
          uint8_t b = uint8_t(val) & 0x7f;
          val >>= 7;
